Add support for es=sf, updated todo, refactor a few things like  EVENTSOURCE_API_CACHE_TTL
