+++ .env.example
@@ -23,7 +23,7 @@ MAPLIBRE_STYLE_URL=https://basemaps.cartocdn.com/gl/voyager-gl-style/style.json
 # leave blank or not set to use default that may vary by file / module.
 LOG_LEVEL=INFO
 
-# how long should server cache events from api call to an eventSource.  Default is 600 secs, aka 10 mins
+# how long should server use cached events from eventSource api call.  Default is 600 secs, aka 10 mins
 # set to 1 to effectively disable caching of eventSource API calls.
-EVENTSOURCE_API_CACHE_EXPIRE_SECS = 600 
+EVENTSOURCE_API_CACHE_TTL = 600
