+++ package.json
@@ -11,6 +11,7 @@
     },
     "dependencies": {
         "@upstash/redis": "^1.25.1",
+        "@vercel/functions": "^2.0.3",
         "@vercel/kv": "^0.2.4",
         "axios": "^1.6.2",
         "date-fns": "^2.30.0",
