+++ src/lib/events/examples.ts
@@ -1,0 +1,21 @@
+// Example event sources for demonstration
+export const exampleEventSources = [
+    {
+        name: 'SF Bay Area Facebook Events (Google Calendar)',
+        id: 'gc:aabe6c219ee2af5b791ea6719e04a92990f9ccd1e68a3ff0d89bacd153a0b36d@group.calendar.google.com',
+        shortId: 'sf', // can just do es=sf in the url
+    },
+    {
+        name: 'Geocaching in Spain (Google Calendar)',
+        id: 'gc:geocachingspain@gmail.com',
+    },
+    {
+        name: 'Protests from pol-rev.com',
+        id: 'protest:all',
+    },
+    // Facebook examples will be added when supported
+    // {
+    //     name: 'Sample Facebook Events',
+    //     id: 'sample-facebook-id',
+    // },
+]
